lib: path: convert max_speed to double instead of int
Using double instead of int allows more precision
to fine tune max_speed parameter.

+++ lib/path/include/path.h
@@ -6,7 +6,7 @@
 typedef struct {
     pose_t pos;
     uint8_t allow_reverse;
-    uint8_t max_speed;
+    double max_speed;
     func_cb_t act;
 } path_pose_t;
